requirements.txt
Flask==3.1.0
edge-tts==6.1.12
google-genai==1.11.0
pillow==11.2.1
pymongo==4.11.3
pyttsx3==2.98
python-dotenv==1.0.1
